Finished changes for schema prefix.  Added ability to read from arbitrary columns for application, experiment, and trial.  Old data getting/setting methods are gone since they are not part of the "core" schema.
git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2772 a8e31b63-e6ca-409b-a01b-2eec1b506579

+++ tools/src/dms/src/edu/uoregon/tau/dms/dss/ParaProfDataSession.java
@@ -413,6 +413,7 @@ public abstract class ParaProfDataSession  extends DataSession{
 	    function.setTotalNumberOfSubRoutines(subrSum);
 
 	    // mean is just the total / numThreads
+
 	    function.setMeanNumberOfCalls((double)callSum / numThreads);
 	    function.setMeanNumberOfSubRoutines((double)subrSum / numThreads);
 
@@ -516,8 +517,8 @@ public abstract class ParaProfDataSession  extends DataSession{
 	double[] exclSum = new double[numberOfMetrics];
 	double[] inclSum = new double[numberOfMetrics];
 	double[] maxInclSum = new double[numberOfMetrics];
-	int callSum = 0;
-	int subrSum = 0;
+	double callSum = 0;
+	double subrSum = 0;
 
 	for (int i=0;i<numberOfMetrics;i++) {
 	    maxInclSum[i] = 0;
